Changed enemy locations from integers to floats Implemented set_x, set_y for enemies

+++ Zombie Tower Defense/EnemyModel.hpp
@@ -22,16 +22,16 @@ public:
     // TODO: some color variable(s) -> R, G, B?
     
     // Location Variables
-    int x;
-    int y;
+    float x;
+    float y;
     
     // Public interface
     
-    int get_x_location();
-    void set_x_location(int x);
+    float get_x_location();
+    void set_x_location(float x);
     
-    int get_y_location();
-    void set_y_location(int y);
+    float get_y_location();
+    void set_y_location(float y);
     
     int get_enemy_health();
     void set_enemy_health(int enemy_health);
